requirements.txt
Django==1.3.1
psycopg2==2.4.2
PIL==1.1.7
oauth==1.0.1
oauth2==1.5.170
python-twitter==0.8.2
simplejson==2.1.6
amqplib==1.0.2
anyjson==0.3.1
celery==2.5.1
django-celery==2.5.1
kombu==2.1.2
django-kombu==0.9.4
django-picklefield==0.2.0
importlib==1.0.2
ordereddict==1.1
python-dateutil==1.5
